requirement.txt
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.1.1
cycler==0.12.1
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.54.1
gdown==5.2.0
h5py==3.11.0
humanfriendly==10.0
idna==3.10
imageio==2.35.1
imgviz==1.7.5
importlib_resources==6.4.5
joblib==1.4.2
kiwisolver==1.4.7
labelme==5.5.0
lazy_loader==0.4
matplotlib==3.7.5
mpmath==1.3.0
natsort==8.4.0
networkx==3.1
nibabel==5.2.1
numpy==1.24.4
onnxruntime==1.19.2
opencv-python==4.10.0.84
packaging==24.1
pillow==10.4.0
protobuf==5.28.2
pyparsing==3.1.4
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
pyreadline3==3.5.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
PyWavelets==1.4.1
PyYAML==6.0.2
QtPy==2.4.1
requests==2.32.3
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
SimpleITK==2.4.0
six==1.16.0
soupsieve==2.6
sympy==1.13.3
termcolor==2.4.0
threadpoolctl==3.5.0
tifffile==2023.7.10
tqdm==4.66.5
urllib3==2.2.3
zipp==3.20.2
